Update FileClass
Made it so that you can read in from a file (only one line works so far). But I have set it so that the user will need to follow all instructions about files from my last commit, but instead of separating by files they will need to have a period at the end of every name, connection, and weight so that it is read in from the file properly.